Chore: casing 오류 해결 #82
